+++ app/src/auth.lua
@@ -21,6 +21,7 @@ local function auth(app)
 	app:match("hide_post", "/post/:post_id[%d]/hide", r2(require("actions.hide")))
 	app:match("saved", "/saved", r2(require("actions.saved")))
 	app:match("mod_remove", "/post/:post_id[%d]/remove", r2(require("actions.mod_remove")))
+	app:match("inbox", "/inbox", r2(require("actions.inbox")))
 
 	-- app:match("prefs", "/prefs", function(self) end) -- stub